+++ package/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-preview.3] - 2018-06-11
+- Add extensions to the XRReferencePointSubsystem
+    - Add support for "attaching" reference points to other trackables, like planes.
+
 ## [1.0.0-preview.2] - 2018-05-23
 Move WIP face subsystem to ignored folder